Playing with integrating webservice into steps
Just messing around with how I could integrate the webservice into the step system. I see a need to have services be able to fire events when http transactions are invoked/completed. This replaces things like "attempt to sync and wait a few seconds" type scenarios with "attempt to sync and once the data has been returned to the device then run a step on the device".